backend/requirements.txt
fastapi
uvicorn
groq
langchain
langchain-groq
duckdb
pandas
requests
beautifulsoup4
python-dotenv